+++ PSModuleDevelopment/changelog.md
@@ -1,6 +1,7 @@
 ﻿# Changelog
 ##
  - New: Export-PSMDString - Parses strings from modules using the PSFramework localization feature.
+ - Upd: Measure-PSMDCommand - Renamed from Measure-PSMDCommandEx, performance upgrades, adding option for comparing multiple test sets.
  - Upd: Refactored and updated the ModuleDebug component
  - Upd: Renamed Get-PSMDHelpEx to Get-PSMDHelp
  - Upd: Template PSFProject - Adding `-IncludAZ` switch parameter to `vsts-packageFunction.ps1`, making the template include the AZ module as managed dependency.